Tracked down this bugger...
I've been hunting this reference bug for 30 minutes.
This has fixed it, finally.
Beware of this in the future: oop::objectify creates some default
methods that won't be overwritten with inherited objects, remember. This
is intended, so if you need to inherit an object's constructors straight
up, nil the one inherited from objectify.